test(mcp): split test_mcp.py into a per-tool package
Break the 1443-line tests/unit/server/test_mcp.py monolith into a
tests/unit/server/mcp/ package: the DB/MCP setup fixtures (engine,
session, fts, vector_store, populated_db, setup_mcp, health_data, _NOW)
move to conftest.py, and tests are grouped per MCP tool (overview,
context+truncation, risk, why, search, dead_code, health) plus
test_config.py for the config/lifespan helpers. Test bodies moved
verbatim; same 42 tests collected and passing.
