+++ tests/unit/server/mcp/test_search.py
@@ -1,0 +1,31 @@
+"""Unit tests for repowise MCP server tools.
+
+Tests all 9 MCP tools using an in-memory SQLite database with pre-populated
+test data, mirroring the conftest pattern from the REST API tests.
+"""
+
+from __future__ import annotations
+
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_search_codebase(setup_mcp):
+    # Index pages in the MCP module's vector store (which is the InMemoryVectorStore)
+    import repowise.server.mcp_server as mcp_mod
+    from repowise.server.mcp_server import search_codebase
+
+    await mcp_mod._vector_store.embed_and_upsert(
+        "file_page:src/auth/service.py",
+        "Auth Service — Main authentication service class",
+        {"title": "Auth Service", "page_type": "file_page", "target_path": "src/auth/service.py"},
+    )
+    await mcp_mod._vector_store.embed_and_upsert(
+        "file_page:src/db/models.py",
+        "DB Models — SQLAlchemy ORM models",
+        {"title": "DB Models", "page_type": "file_page", "target_path": "src/db/models.py"},
+    )
+
+    result = await search_codebase("authentication service")
+    assert "results" in result
+    assert len(result["results"]) >= 1
